Add files via upload
Prefabs created:

Player2 prefab that can be switched to based on Player1 input. Keeps same same score as Player1 and can switch inbetween.
Hazards - vertical and horizontal - resets player1 and player2 back to start. Can be reworked so that the reset is done by the players instead of these
Vertical Platform that goes up and down between start and a chosen object (recommend to use an empty for this)


Player1 has a few changes compared to original:
-added jump
-added switch to switch between player2 and player1 using "m" and "l"
--Problem with using same key as it doesn't switch back to player1 if they use the same key to switch